Rename trackMixpanel file
