+++ apps/waitlist/lib/mixpanel/trackWaitlistMixpanelEvent.ts
